Fixed bugs with the apl ignoring good vivify casts. The lookahead system was too buggy for targets healed, added a custom condition in the mw folder that utilizes event links instead
