+++ src/analysis/retail/monk/mistweaver/CombatLogParser.ts
@@ -81,7 +81,7 @@ import Guide from './Guide';
 import SheilunsGiftCloudTracker from './modules/spells/SheilunsGiftCloudTracker';
 import SheilunsGiftCloudGraph from './modules/spells/SheilunsGiftCloudGraph';
 import HotCountGraph from './modules/features/HotCountGraph';
-import AplCheck from './modules/core/AplCheck';
+import AplCheck from './modules/core/apl/AplCheck';
 
 class CombatLogParser extends CoreCombatLogParser {
   static specModules = {
