+++ src/analysis/retail/monk/mistweaver/modules/core/apl/ExplainCelestial.tsx
@@ -2,7 +2,7 @@ import Spell from 'common/SPELLS/Spell';
 import { useAnalyzer } from 'interface/guide';
 import { ViolationExplainer } from 'interface/guide/components/Apl/violations/claims';
 import { InternalRule } from 'parser/shared/metrics/apl';
-import BaseCelestialAnalyzer from '../spells/BaseCelestialAnalyzer';
+import BaseCelestialAnalyzer from '../../spells/BaseCelestialAnalyzer';
 
 export const filterCelestial = (
   explainer: ViolationExplainer<{ rule: InternalRule; spell: Spell }>,
